Refactor: Move `Master*` models from `Goods` to `Masters`
