+++ app/Actions/Helpers/Media/SaveModelImages.php
@@ -11,8 +11,8 @@ namespace App\Actions\Helpers\Media;
 use App\Actions\Helpers\Media\Hydrators\MediaHydrateMultiplicity;
 use App\Actions\Helpers\Media\Hydrators\MediaHydrateUsage;
 use App\Models\Catalogue\Product;
-use App\Models\Goods\MasterAsset;
 use App\Models\Helpers\Media;
+use App\Models\Masters\MasterAsset;
 use App\Models\Web\WebBlock;
 use Lorisleiva\Actions\Concerns\AsAction;
 use stdClass;
