+++ app/Actions/Masters/MasterAsset/UI/IndexMasterAssets.php
@@ -14,9 +14,9 @@ use App\Actions\GrpAction;
 use App\Actions\Masters\MasterShop\UI\ShowMasterShop;
 use App\Http\Resources\Goods\Catalogue\MasterProductsResource;
 use App\InertiaTable\InertiaTable;
-use App\Models\Goods\MasterAsset;
-use App\Models\Goods\MasterProductCategory;
-use App\Models\Goods\MasterShop;
+use App\Models\Masters\MasterAsset;
+use App\Models\Masters\MasterProductCategory;
+use App\Models\Masters\MasterShop;
 use App\Models\SysAdmin\Group;
 use App\Services\QueryBuilder;
 use Illuminate\Contracts\Pagination\LengthAwarePaginator;
