+++ app/Actions/Masters/MasterAsset/UpdateMasterAsset.php
@@ -13,8 +13,8 @@ use App\Actions\SysAdmin\Group\Hydrators\GroupHydrateMasterAssets;
 use App\Actions\Traits\Rules\WithNoStrictRules;
 use App\Actions\Traits\WithActionUpdate;
 use App\Enums\Catalogue\MasterProductCategory\MasterProductCategoryTypeEnum;
-use App\Models\Goods\MasterAsset;
-use App\Models\Goods\MasterProductCategory;
+use App\Models\Masters\MasterAsset;
+use App\Models\Masters\MasterProductCategory;
 use App\Rules\AlphaDashDot;
 use App\Rules\IUnique;
 use Illuminate\Support\Arr;
