+++ app/Actions/Masters/MasterProductCategory/StoreMasterProductCategory.php
@@ -15,8 +15,8 @@ use App\Actions\Traits\Rules\WithNoStrictRules;
 use App\Enums\Catalogue\MasterProductCategory\MasterProductCategoryTypeEnum;
 use App\Enums\Catalogue\ProductCategory\ProductCategoryTypeEnum;
 use App\Enums\Helpers\TimeSeries\TimeSeriesFrequencyEnum;
-use App\Models\Goods\MasterProductCategory;
-use App\Models\Goods\MasterShop;
+use App\Models\Masters\MasterProductCategory;
+use App\Models\Masters\MasterShop;
 use App\Rules\AlphaDashDot;
 use App\Rules\IUnique;
 use Illuminate\Validation\Rule;
