+++ app/Actions/Masters/MasterProductCategory/UI/IndexMasterFamilies.php
@@ -15,8 +15,8 @@ use App\Actions\OrgAction;
 use App\Enums\Catalogue\ProductCategory\ProductCategoryTypeEnum;
 use App\Http\Resources\Goods\Catalogue\MasterFamiliesResource;
 use App\InertiaTable\InertiaTable;
-use App\Models\Goods\MasterProductCategory;
-use App\Models\Goods\MasterShop;
+use App\Models\Masters\MasterProductCategory;
+use App\Models\Masters\MasterShop;
 use App\Models\SysAdmin\Group;
 use App\Services\QueryBuilder;
 use Closure;
