+++ app/Actions/Masters/MasterProductCategory/UI/IndexMasterSubDepartments.php
@@ -15,8 +15,8 @@ use App\Actions\Masters\MasterShop\UI\ShowMasterShop;
 use App\Enums\Catalogue\ProductCategory\ProductCategoryTypeEnum;
 use App\Http\Resources\Goods\Catalogue\MasterSubDepartmentsResource;
 use App\InertiaTable\InertiaTable;
-use App\Models\Goods\MasterProductCategory;
-use App\Models\Goods\MasterShop;
+use App\Models\Masters\MasterProductCategory;
+use App\Models\Masters\MasterShop;
 use App\Services\QueryBuilder;
 use Closure;
 use Illuminate\Contracts\Pagination\LengthAwarePaginator;
