+++ app/Actions/Masters/MasterProductCategory/UI/ShowMasterFamily.php
@@ -16,8 +16,8 @@ use App\Actions\Comms\Mailshot\UI\IndexMailshots;
 use App\Actions\GrpAction;
 use App\Enums\UI\SupplyChain\MasterFamilyTabsEnum;
 use App\Http\Resources\Catalogue\DepartmentsResource;
-use App\Models\Goods\MasterProductCategory;
-use App\Models\Goods\MasterShop;
+use App\Models\Masters\MasterProductCategory;
+use App\Models\Masters\MasterShop;
 use Inertia\Inertia;
 use Inertia\Response;
 use Lorisleiva\Actions\ActionRequest;
