+++ app/Actions/Masters/MasterProductCategory/UI/ShowMasterFamilyWorkshop.php
@@ -19,8 +19,8 @@ use App\Http\Resources\Catalogue\DepartmentsResource;
 use App\Http\Resources\Catalogue\FamilyResource;
 use App\Http\Resources\Goods\Catalogue\MasterProductsResource;
 use App\Http\Resources\Web\WebBlockTypesResource;
-use App\Models\Goods\MasterProductCategory;
-use App\Models\Goods\MasterShop;
+use App\Models\Masters\MasterProductCategory;
+use App\Models\Masters\MasterShop;
 use App\Models\Web\WebBlockType;
 use Inertia\Inertia;
 use Inertia\Response;
