+++ app/Actions/Masters/MasterProductCategory/UI/ShowMasterSubDepartment.php
@@ -17,8 +17,8 @@ use App\Actions\Helpers\History\UI\IndexHistory;
 use App\Enums\UI\SupplyChain\MasterSubDepartmentTabsEnum;
 use App\Http\Resources\Catalogue\DepartmentsResource;
 use App\Http\Resources\History\HistoryResource;
-use App\Models\Goods\MasterProductCategory;
-use App\Models\Goods\MasterShop;
+use App\Models\Masters\MasterProductCategory;
+use App\Models\Masters\MasterShop;
 use Inertia\Inertia;
 use Inertia\Response;
 use Lorisleiva\Actions\ActionRequest;
