+++ app/Actions/Masters/MasterShop/HydrateMasterShop.php
@@ -11,7 +11,7 @@ namespace App\Actions\Masters\MasterShop;
 use App\Actions\HydrateModel;
 use App\Actions\Masters\MasterShop\Hydrators\MasterShopHydrateShops;
 use App\Actions\Traits\WithNormalise;
-use App\Models\Goods\MasterShop;
+use App\Models\Masters\MasterShop;
 use Illuminate\Console\Command;
 use Illuminate\Support\Collection;
