+++ app/Actions/Traits/WithAttachMediaToModel.php
@@ -11,8 +11,8 @@ namespace App\Actions\Traits;
 use App\Models\Catalogue\Product;
 use App\Models\Catalogue\Shop;
 use App\Models\CRM\WebUser;
-use App\Models\Goods\MasterProductCategory;
 use App\Models\Helpers\Media;
+use App\Models\Masters\MasterProductCategory;
 use App\Models\SysAdmin\Group;
 use App\Models\SysAdmin\Organisation;
 use App\Models\SysAdmin\User;
