+++ app/Actions/Transfers/Aurora/FetchAuroraMasterFamilies.php
@@ -10,7 +10,7 @@ namespace App\Actions\Transfers\Aurora;
 
 use App\Actions\Masters\MasterProductCategory\StoreMasterProductCategory;
 use App\Actions\Masters\MasterProductCategory\UpdateMasterProductCategory;
-use App\Models\Goods\MasterProductCategory;
+use App\Models\Masters\MasterProductCategory;
 use App\Transfers\SourceOrganisationService;
 use Exception;
 use Illuminate\Database\Query\Builder;
