+++ app/Models/Catalogue/Shop.php
@@ -40,7 +40,6 @@ use App\Models\Dispatching\Picking;
 use App\Models\Dropshipping\CustomerClient;
 use App\Models\Dropshipping\Portfolio;
 use App\Models\Fulfilment\Fulfilment;
-use App\Models\Goods\MasterShop;
 use App\Models\Helpers\Address;
 use App\Models\Helpers\Country;
 use App\Models\Helpers\Currency;
@@ -52,6 +51,7 @@ use App\Models\Helpers\TaxNumber;
 use App\Models\Helpers\Timezone;
 use App\Models\Helpers\UniversalSearch;
 use App\Models\Helpers\Upload;
+use App\Models\Masters\MasterShop;
 use App\Models\Ordering\Adjustment;
 use App\Models\Ordering\Order;
 use App\Models\Ordering\Purge;
