+++ app/Models/Masters/MasterAssetOrderingIntervals.php
@@ -2,21 +2,19 @@
 
 /*
  * Author: Raul Perusquia <raul@inikoo.com>
- * Created: Sat, 28 Dec 2024 18:00:26 Malaysia Time, Kuala Lumpur, Malaysia
+ * Created: Mon, 30 Dec 2024 12:48:10 Malaysia Time, Kuala Lumpur, Malaysia
  * Copyright (c) 2024, Raul A Perusquia Flores
  */
 
-namespace App\Models\Goods;
+namespace App\Models\Masters;
 
-use App\Models\Catalogue\Shop;
 use Illuminate\Database\Eloquent\Model;
-use Illuminate\Database\Eloquent\Relations\BelongsTo;
 
 /**
  *
  *
  * @property int $id
- * @property int $master_shop_id
+ * @property int $master_asset_id
  * @property int $invoices_all
  * @property int $invoices_1y
  * @property int $invoices_1q
@@ -173,45 +171,6 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
  * @property int $delivery_notes_pq3
  * @property int $delivery_notes_pq4
  * @property int $delivery_notes_pq5
- * @property int $registrations_all
- * @property int $registrations_1y
- * @property int $registrations_1q
- * @property int $registrations_1m
- * @property int $registrations_1w
- * @property int $registrations_3d
- * @property int $registrations_1d
- * @property int $registrations_ytd
- * @property int $registrations_qtd
- * @property int $registrations_mtd
- * @property int $registrations_wtd
- * @property int $registrations_tdy
- * @property int $registrations_lm
- * @property int $registrations_lw
- * @property int $registrations_ld
- * @property int $registrations_1y_ly
- * @property int $registrations_1q_ly
- * @property int $registrations_1m_ly
- * @property int $registrations_1w_ly
- * @property int $registrations_3d_ly
- * @property int $registrations_1d_ly
- * @property int $registrations_ytd_ly
- * @property int $registrations_qtd_ly
- * @property int $registrations_mtd_ly
- * @property int $registrations_wtd_ly
- * @property int $registrations_tdy_ly
- * @property int $registrations_lm_ly
- * @property int $registrations_lw_ly
- * @property int $registrations_ld_ly
- * @property int $registrations_py1
- * @property int $registrations_py2
- * @property int $registrations_py3
- * @property int $registrations_py4
- * @property int $registrations_py5
- * @property int $registrations_pq1
- * @property int $registrations_pq2
- * @property int $registrations_pq3
- * @property int $registrations_pq4
- * @property int $registrations_pq5
  * @property int $customers_invoiced_all
  * @property int $customers_invoiced_1y
  * @property int $customers_invoiced_1q
@@ -253,19 +212,14 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
  * @property int $customers_invoiced_pq5
  * @property \Illuminate\Support\Carbon|null $created_at
  * @property \Illuminate\Support\Carbon|null $updated_at
- * @property-read \App\Models\Goods\MasterShop $masterShop
- * @method static \Illuminate\Database\Eloquent\Builder<static>|MasterShopOrderingIntervals newModelQuery()
- * @method static \Illuminate\Database\Eloquent\Builder<static>|MasterShopOrderingIntervals newQuery()
- * @method static \Illuminate\Database\Eloquent\Builder<static>|MasterShopOrderingIntervals query()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|MasterAssetOrderingIntervals newModelQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|MasterAssetOrderingIntervals newQuery()
+ * @method static \Illuminate\Database\Eloquent\Builder<static>|MasterAssetOrderingIntervals query()
  * @mixin \Eloquent
  */
-class MasterShopOrderingIntervals extends Model
+class MasterAssetOrderingIntervals extends Model
 {
-    protected $table = 'master_shop_ordering_intervals';
-
+    protected $table = 'master_asset_ordering_intervals';
     protected $guarded = [];
-    public function masterShop(): BelongsTo
-    {
-        return $this->belongsTo(MasterShop::class);
-    }
+
 }
