+++ app/Models/Masters/MasterAssetStats.php
@@ -6,7 +6,7 @@
  * Copyright (c) 2024, Raul A Perusquia Flores
  */
 
-namespace App\Models\Goods;
+namespace App\Models\Masters;
 
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
@@ -36,7 +36,7 @@ use Illuminate\Database\Eloquent\Relations\BelongsTo;
  * @property int $number_assets_type_shipping_zone
  * @property \Illuminate\Support\Carbon|null $created_at
  * @property \Illuminate\Support\Carbon|null $updated_at
- * @property-read \App\Models\Goods\MasterAsset $masterAsset
+ * @property-read \App\Models\Masters\MasterAsset $masterAsset
  * @method static \Illuminate\Database\Eloquent\Builder<static>|MasterAssetStats newModelQuery()
  * @method static \Illuminate\Database\Eloquent\Builder<static>|MasterAssetStats newQuery()
  * @method static \Illuminate\Database\Eloquent\Builder<static>|MasterAssetStats query()
