+++ app/Models/SysAdmin/Group.php
@@ -55,9 +55,6 @@ use App\Models\Fulfilment\FulfilmentCustomer;
 use App\Models\Fulfilment\RecurringBill;
 use App\Models\Fulfilment\Space;
 use App\Models\Goods\Ingredient;
-use App\Models\Goods\MasterAsset;
-use App\Models\Goods\MasterProductCategory;
-use App\Models\Goods\MasterShop;
 use App\Models\Goods\Stock;
 use App\Models\Goods\StockFamily;
 use App\Models\Goods\TradeUnit;
@@ -71,6 +68,9 @@ use App\Models\HumanResources\JobPosition;
 use App\Models\Inventory\Location;
 use App\Models\Inventory\Warehouse;
 use App\Models\Inventory\WarehouseArea;
+use App\Models\Masters\MasterAsset;
+use App\Models\Masters\MasterProductCategory;
+use App\Models\Masters\MasterShop;
 use App\Models\Ordering\Adjustment;
 use App\Models\Ordering\Order;
 use App\Models\Ordering\Purge;
