+++ app/Transfers/Aurora/WithAuroraParsers.php
@@ -96,7 +96,6 @@ use App\Models\Discounts\OfferComponent;
 use App\Models\Dispatching\DeliveryNote;
 use App\Models\Dispatching\Shipper;
 use App\Models\Goods\Ingredient;
-use App\Models\Goods\MasterProductCategory;
 use App\Models\Goods\Stock;
 use App\Models\Goods\TradeUnit;
 use App\Models\Helpers\Barcode;
@@ -113,6 +112,7 @@ use App\Models\Inventory\Location;
 use App\Models\Inventory\OrgStock;
 use App\Models\Inventory\Warehouse;
 use App\Models\Inventory\WarehouseArea;
+use App\Models\Masters\MasterProductCategory;
 use App\Models\Ordering\Adjustment;
 use App\Models\Ordering\Order;
 use App\Models\Ordering\Purge;
