+++ tests/Feature/AikuSections/GoodsTest.php
@@ -31,24 +31,24 @@ use App\Enums\Goods\MasterAsset\MasterAssetTypeEnum;
 use App\Enums\Goods\Stock\StockStateEnum;
 use App\Enums\Goods\StockFamily\StockFamilyStateEnum;
 use App\Models\Goods\Ingredient;
-use App\Models\Goods\MasterAsset;
-use App\Models\Goods\MasterAssetOrderingIntervals;
-use App\Models\Goods\MasterAssetOrderingStats;
-use App\Models\Goods\MasterAssetSalesIntervals;
-use App\Models\Goods\MasterAssetStats;
-use App\Models\Goods\MasterProductCategory;
-use App\Models\Goods\MasterProductCategoryOrderingIntervals;
-use App\Models\Goods\MasterProductCategoryOrderingStats;
-use App\Models\Goods\MasterProductCategorySalesIntervals;
-use App\Models\Goods\MasterProductCategoryStats;
-use App\Models\Goods\MasterShop;
-use App\Models\Goods\MasterShopOrderingIntervals;
-use App\Models\Goods\MasterShopOrderingStats;
-use App\Models\Goods\MasterShopSalesIntervals;
-use App\Models\Goods\MasterShopStats;
 use App\Models\Goods\Stock;
 use App\Models\Goods\StockFamily;
 use App\Models\Goods\TradeUnit;
+use App\Models\Masters\MasterAsset;
+use App\Models\Masters\MasterAssetOrderingIntervals;
+use App\Models\Masters\MasterAssetOrderingStats;
+use App\Models\Masters\MasterAssetSalesIntervals;
+use App\Models\Masters\MasterAssetStats;
+use App\Models\Masters\MasterProductCategory;
+use App\Models\Masters\MasterProductCategoryOrderingIntervals;
+use App\Models\Masters\MasterProductCategoryOrderingStats;
+use App\Models\Masters\MasterProductCategorySalesIntervals;
+use App\Models\Masters\MasterProductCategoryStats;
+use App\Models\Masters\MasterShop;
+use App\Models\Masters\MasterShopOrderingIntervals;
+use App\Models\Masters\MasterShopOrderingStats;
+use App\Models\Masters\MasterShopSalesIntervals;
+use App\Models\Masters\MasterShopStats;
 use Inertia\Testing\AssertableInertia;
 
 use function Pest\Laravel\actingAs;
